+++ kinto/tests/core/resource/test_model.py
@@ -1,5 +1,3 @@
-import mock
-
 from pyramid import httpexceptions
 
 from kinto.tests.core.resource import BaseTest
@@ -9,11 +7,6 @@ class ModelTest(BaseTest):
     def setUp(self):
         super(ModelTest, self).setUp()
         self.record = self.model.create_record({'field': 'value'})
-
-    def test_access_to_collection_is_deprecated(self):
-        with mock.patch('kinto.core.resource.warnings.warn') as mocked:
-            self.resource.collection
-            self.assertTrue(mocked.called)
 
     def test_list_gives_number_of_results_in_headers(self):
         self.resource.collection_get()